requirements.txt
annotated-types==0.7.0
anyio==4.3.0
astor==0.8.1
attrdict==2.0.1
Babel==2.15.0
bce-python-sdk==0.9.11
bcrypt==4.1.3
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.7
cssselect==1.2.0
cssutils==2.11.0
cycler==0.12.1
Cython==3.0.10
decorator==5.1.1
dnspython==2.6.1
email_validator==2.1.1
et-xmlfile==1.1.0
fastapi==0.111.0
fastapi-cli==0.0.4
fire==0.6.0
Flask==3.0.3
flask-babel==4.0.0
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-MySQLdb==2.0.0
Flask-SocketIO==5.3.6
Flask-WTF==1.2.1
fonttools==4.51.0
future==1.0.0
google-ai-generativelanguage
google-api-core
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai
googleapis-common-protos
grpcio==1.64.0
grpcio-status
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.7
imageio==2.34.1
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jose==1.0.0
jsonpickle==3.0.4
jwt==1.3.1
kiwisolver==1.4.5
lazy_loader==0.4
lmdb==1.4.1
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mysql-connector-python==8.4.0
mysqlclient==2.2.4
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.10.3
packaging==24.0
paddleocr==2.7.3
paddlepaddle==2.6.1
pandas==2.2.2
pdf2docx==0.5.8
pdf2image==1.17.0
pillow==10.3.0
premailer==3.10.0
proto-plus==1.23.0
protobuf==3.20.2
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.18.0
PyMuPDF==1.21.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.2
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.1
rarfile==4.2
requests==2.32.2
rich==13.7.1
rsa==4.9
scikit-image==0.23.2
scipy==1.13.1
shapely==2.0.4
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
termcolor==2.4.0
tifffile==2024.5.22
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
visualdl==2.5.3
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.3
wsproto==1.2.0
WTForms==3.1.2
